wip(refactor): Attempted refactor, have new idea coming.
