+++ interface/src/interface.rs
@@ -2,10 +2,8 @@ use crate::{
     iface::{
         interface_service_client::InterfaceServiceClient, ClientListRequest, ConnectionRequest, AddTaskRequest,
     },
-    ui::{App, UI, Action}, common::ShellCode,
+    tui::{Tui, Result, Action, App}, common::ShellCode,
 };
-
-use std::error::Error;
 
 use tonic::transport::Channel;
 
@@ -15,27 +13,25 @@ pub struct Interface {
     client: InterfaceServiceClient<Channel>,
     uuid: String,
 
-    ui: UI,
+    ui: Tui,
 }
 
 impl Interface {
-    pub async fn connect(url: &'static str) -> Result<Self, Box<dyn Error>> {
+    pub async fn connect(url: &'static str) -> Result<Self> {
         let mut client = InterfaceServiceClient::connect(url).await?;
         let response = client
             .connection(tonic::Request::new(ConnectionRequest {}))
             .await?
             .into_inner();
 
-        // println!("Connect: RESPONSE={:?}", response);
-
         Ok(Self {
             client,
             uuid: response.uuid,
-            ui: UI::new()?,
+            ui: Tui::new()?,
         })
     }
 
-    async fn get_list(&mut self) -> Result<Vec<String>, Box<dyn Error>> {
+    async fn get_list(&mut self) -> Result<Vec<String>> {
         let response = self
             .client
             .get_list(tonic::Request::new(ClientListRequest {
@@ -44,12 +40,10 @@ impl Interface {
             .await?
             .into_inner();
 
-        // println!("Poll: RESPONSE={:?}", response);
-
         Ok(response.list)
     }
 
-    async fn add_task(&mut self, client_uuid: String, task: String) -> Result<(), Box<dyn Error>> {
+    async fn add_task(&mut self, client_uuid: String, task: String) -> Result<()> {
         let _ = self
             .client
             .add_task(tonic::Request::new(AddTaskRequest {
@@ -63,12 +57,11 @@ impl Interface {
         Ok(())
     }
 
-    pub async fn run(mut self) -> Result<(), Box<dyn Error>> {
+    pub async fn run(mut self) -> Result<()> {
         let mut app = App::new();
         app.update_clients(self.get_list().await?);
 
         loop {
-            // let list = Drawable::ClientList(self.get_list().await?);
             if let Some(action) = self.ui.events(&mut app, 250)? {
                 match action {
                     Action::SendTask(u, t) => self.add_task(u, t).await?,
